requirements.txt
flask
aws-lambda-wsgi
python-dotenv
pymongo
pytest
flasgger